no usar __init__
